refactor(container): extract Google OAuth provider definitions
